requirements.txt
Django==3.2
djangorestframework==3.12.4
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.1.3
Flask-API==3.0.post1
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
requests==2.28.1
urllib3==1.26.13
Werkzeug==2.1.2
zipp==3.8.1
PyJWT==1.7.1